Added a very hacky multi-Spark fetcher implementation
The idea is to allow configuring multiple fetcher for the same application
type with an implicit assumption that for any application only a single fetcher
could yield results.

Example configuration

  <fetcher>
    <applicationtype>spark</applicationtype>
    <classname>com.linkedin.drelephant.spark.fetchers.SparkFetcher</classname>
    <params>
      <use_rest_for_eventlogs>true</use_rest_for_eventlogs>
      <spark_conf_dir>/usr/lib/spark/2.1/conf</spark_conf_dir>
    </params>
  </fetcher>

  <fetcher>
    <applicationtype>spark</applicationtype>
    <classname>com.linkedin.drelephant.spark.fetchers.SparkFetcher</classname>
    <params>
      <use_rest_for_eventlogs>true</use_rest_for_eventlogs>
      <spark_conf_dir>/usr/lib/spark/1.6/conf</spark_conf_dir>
    </params>
  </fetcher>

Once a new application arrives, all fetchers would be asked to fetch it
with at most one succeeding. If all of them fail, the analysis would be
rescheduled.

See #226 for context and motivation for this change.